debugging merge stack, not workin

+++ src/lib/user/malloc.h
@@ -1,7 +1,7 @@
 #ifndef __LIB_USER_MALLOC_H
 #define __LIB_USER_MALLOC_H
 
-void * malloc(size_t size);
+void * malloc(int size);
 void free(void *ptr);
 
 #endif
